move the ServerInfo to the settings page
